feat: add support for machine client audience
This commit adds a new optional property `machineClientAudience` to the
`IAuth` interface. This property allows specifying the audience value to
be used when generating access tokens for the machine client using the
Client Credentials flow.

Additionally, the README has been updated to include links to the
Serverless Reference Architecture and the SBT Workshop.